trade_log_parser/requirements.txt
pandas
pillow
pytesseract
opencv-python
anthropic
openai